Move some code around to avoid import cycles
