+++ common/dynamicconfig/client.go
@@ -100,3 +100,7 @@ type (
 		TaskType      enumsspb.TaskType
 	}
 )
+
+func (k Key) String() string {
+	return string(k)
+}
