+++ common/persistence/cassandra/test.go
@@ -30,7 +30,6 @@ import (
 	"time"
 
 	"github.com/gocql/gocql"
-	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/backoff"
 	"go.temporal.io/server/common/config"
 	"go.temporal.io/server/common/debug"
@@ -40,6 +39,7 @@ import (
 	"go.temporal.io/server/common/metrics"
 	p "go.temporal.io/server/common/persistence"
 	commongocql "go.temporal.io/server/common/persistence/nosql/nosqlplugin/cassandra/gocql"
+	"go.temporal.io/server/common/primitives"
 	"go.temporal.io/server/common/resolver"
 	"go.temporal.io/server/environment"
 	"go.temporal.io/server/tests/testutils"
@@ -91,11 +91,11 @@ func NewTestCluster(keyspace, username, password, host string, port int, schemaD
 func (s *TestCluster) Config() config.Persistence {
 	cfg := s.cfg
 	return config.Persistence{
-		DefaultStore:    "test",
+		DefaultStore: "test",
 		DataStores: map[string]config.DataStore{
 			"test": {Cassandra: &cfg, FaultInjection: s.faultInjection},
 		},
-		TransactionSizeLimit: dynamicconfig.GetIntPropertyFn(common.DefaultTransactionSizeLimit),
+		TransactionSizeLimit: dynamicconfig.GetIntPropertyFn(primitives.DefaultTransactionSizeLimit),
 	}
 }
