+++ common/util_test.go
@@ -158,117 +158,6 @@ func TestValidateRetryPolicy(t *testing.T) {
 	}
 }
 
-func TestEnsureRetryPolicyDefaults(t *testing.T) {
-	defaultRetrySettings := DefaultRetrySettings{
-		InitialInterval:            time.Second,
-		MaximumIntervalCoefficient: 100,
-		BackoffCoefficient:         2.0,
-		MaximumAttempts:            120,
-	}
-
-	defaultRetryPolicy := &commonpb.RetryPolicy{
-		InitialInterval:    durationpb.New(1 * time.Second),
-		MaximumInterval:    durationpb.New(100 * time.Second),
-		BackoffCoefficient: 2.0,
-		MaximumAttempts:    120,
-	}
-
-	testCases := []struct {
-		name  string
-		input *commonpb.RetryPolicy
-		want  *commonpb.RetryPolicy
-	}{
-		{
-			name:  "default fields are set ",
-			input: &commonpb.RetryPolicy{},
-			want:  defaultRetryPolicy,
-		},
-		{
-			name: "non-default InitialIntervalInSeconds is not set",
-			input: &commonpb.RetryPolicy{
-				InitialInterval: durationpb.New(2 * time.Second),
-			},
-			want: &commonpb.RetryPolicy{
-				InitialInterval:    durationpb.New(2 * time.Second),
-				MaximumInterval:    durationpb.New(200 * time.Second),
-				BackoffCoefficient: 2,
-				MaximumAttempts:    120,
-			},
-		},
-		{
-			name: "non-default MaximumIntervalInSeconds is not set",
-			input: &commonpb.RetryPolicy{
-				MaximumInterval: durationpb.New(1000 * time.Second),
-			},
-			want: &commonpb.RetryPolicy{
-				InitialInterval:    durationpb.New(1 * time.Second),
-				MaximumInterval:    durationpb.New(1000 * time.Second),
-				BackoffCoefficient: 2,
-				MaximumAttempts:    120,
-			},
-		},
-		{
-			name: "non-default BackoffCoefficient is not set",
-			input: &commonpb.RetryPolicy{
-				BackoffCoefficient: 1.5,
-			},
-			want: &commonpb.RetryPolicy{
-				InitialInterval:    durationpb.New(1 * time.Second),
-				MaximumInterval:    durationpb.New(100 * time.Second),
-				BackoffCoefficient: 1.5,
-				MaximumAttempts:    120,
-			},
-		},
-		{
-			name: "non-default Maximum attempts is not set",
-			input: &commonpb.RetryPolicy{
-				MaximumAttempts: 49,
-			},
-			want: &commonpb.RetryPolicy{
-				InitialInterval:    durationpb.New(1 * time.Second),
-				MaximumInterval:    durationpb.New(100 * time.Second),
-				BackoffCoefficient: 2,
-				MaximumAttempts:    49,
-			},
-		},
-		{
-			name: "non-retryable errors are set",
-			input: &commonpb.RetryPolicy{
-				NonRetryableErrorTypes: []string{"testFailureType"},
-			},
-			want: &commonpb.RetryPolicy{
-				InitialInterval:        durationpb.New(1 * time.Second),
-				MaximumInterval:        durationpb.New(100 * time.Second),
-				BackoffCoefficient:     2.0,
-				MaximumAttempts:        120,
-				NonRetryableErrorTypes: []string{"testFailureType"},
-			},
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			EnsureRetryPolicyDefaults(tt.input, defaultRetrySettings)
-			assert.Equal(t, tt.want, tt.input)
-		})
-	}
-}
-
-func Test_FromConfigToRetryPolicy(t *testing.T) {
-	options := map[string]interface{}{
-		initialIntervalInSecondsConfigKey:   2,
-		maximumIntervalCoefficientConfigKey: 100.0,
-		backoffCoefficientConfigKey:         4.0,
-		maximumAttemptsConfigKey:            5,
-	}
-
-	defaultSettings := FromConfigToDefaultRetrySettings(options)
-	assert.Equal(t, 2*time.Second, defaultSettings.InitialInterval)
-	assert.Equal(t, 100.0, defaultSettings.MaximumIntervalCoefficient)
-	assert.Equal(t, 4.0, defaultSettings.BackoffCoefficient)
-	assert.Equal(t, int32(5), defaultSettings.MaximumAttempts)
-}
-
 func TestIsContextDeadlineExceededErr(t *testing.T) {
 	ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(-time.Millisecond))
 	defer cancel()
