+++ tools/sql/clitest/version_tests.go
@@ -31,8 +31,8 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"go.temporal.io/server/common/config"
+	"go.temporal.io/server/common/primitives"
 
-	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/dynamicconfig"
 	persistencetests "go.temporal.io/server/common/persistence/persistence-tests"
 	persistencesql "go.temporal.io/server/common/persistence/sql"
@@ -130,7 +130,7 @@ func (s *VersionTestSuite) TestVerifyCompatibleVersion() {
 			"default":    {SQL: &defaultCfg},
 			"visibility": {SQL: &visibilityCfg},
 		},
-		TransactionSizeLimit: dynamicconfig.GetIntPropertyFn(common.DefaultTransactionSizeLimit),
+		TransactionSizeLimit: dynamicconfig.GetIntPropertyFn(primitives.DefaultTransactionSizeLimit),
 	}
 	s.NoError(persistencesql.VerifyCompatibleVersion(cfg, resolver.NewNoopResolver()))
 }
